Send check results in bulk: report most severe status
A check may monitor many metrics.  Report the most severe state of any
of these metrics as status and attach detailed information on which
metrics are in which state.

For fast access of metrics by status, we cache them by category (OK,
WARNING, CRITICAL) and update the cache incrementally when receiving new
values.