Delay re-calling timeout callback after deadline miss
If we ever miss the deadline for a timeout check, we immediately
(without waiting) call the timeout callback.  To avoid calling it again
the next iteration, throttle the timeout check and wait at least one
timeout period for new values.  If no values arrive within this period,
call the timeout callback again.