Add crude heartbeat timer to periodically send check results
If a metric times out, its status is only resent after the timeout
interval has passed once more, ignoring the (global) resend interval.
This makes the NSCA host think this passive check is defective,
resulting in the state repeatedly entering a CRITICAL state (caused by
the timeout check) and then an UNKNOWN state (caused by the NSCA host
thinking this passive check is defective), only to be toggled back to
CRITICAL once the timeout check decides to resend that state.