fix(state_cache): Optionally ignore invalid metric values
If invalid data points arrive (for example with a non-monotonic
timestamp) and the configuration key `ignore_update_errors` is set to
`true`, the data point is ignored and dropped.

This suppresses the *CRITICAL* report that would be sent otherwise,
while still triggering a timeout-report if no valid data point is
received soon after (=within the timeout period starting from the last
valid data point from).

This implements the 'make ignoring invalid metric values'-part of #15
configurable.