feat(reporter): Wait for outdated checks to stop
Each `Check` has a heartbeat task and optionally some timeout tasks.
Both take at least one iteration of the async loop to handle their own
cancellation, so we asynchronously wait for them to finish properly.

In the case that some task takes too long (>1s) to stop, we raise a
`TimeoutError` and exit.