feat(reporter): Add global metric overrides
Sometimes, a metric is known to temporarily produce "bad" data. We would
like to ignore it from any checks without having to remove it from the
check configuration (as that usually results in misconfigurations when
re-adding the metric).

This change introduces the new top-level configuration key "overrides".
It can be used to define an array of metrics to ignore:

{
  "overrides": {
    "ignored_metrics": [
      "foo.bar.*",
      "qux.baz"
    ]
  }
}

If a check defines a metric that matches any of the patterns in
`ignored_metrics`, values of that metric will be ignored completely.